Fixed %age change output as "-nan%"
This was due to a divide by zero which could be triggered if there is no
difference between the test results across all samples (e.g. they are all
zero to start with).

Also tweaked to supress the deprecated function warning generated from within
scipy when using its ttest_ind() function. It is assumed that if/when the
scipy.stats.mean() function is removed, they will recode their ttest functions
to address this so we won't need any further changes here.